mcp-server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
atlassian-python-api==3.41.1
boto3==1.28.53
pydantic==2.3.0
loguru==0.7.0
jsonrpcserver==5.0.9
markdown==3.5
bs4==0.0.1
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
